% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_groups.R
\name{bq_groups}
\alias{bq_groups}
\title{Briq user groups}
\usage{
bq_groups(max_results_per_page = 100,
  organization = Sys.getenv("organization_name"),
  api_token = Sys.getenv("briq_api_token"))
}
\arguments{
\item{max_results_per_page}{Maximum number of results per page}

\item{organization}{Name of your Briq organization}

\item{api_token}{Briq API token}
}
\value{
Returns a tibble of Briq user groups of your organization
}
\description{
List all the Briq user groups of your organization
}
