% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_revert_transaction.R
\name{bq_revert_transaction}
\alias{bq_revert_transaction}
\title{Revert Briq transaction}
\usage{
bq_revert_transaction(transaction_id,
  organization = Sys.getenv("organization_name"),
  api_token = Sys.getenv("briq_api_token"))
}
\arguments{
\item{transaction_id}{ID of transaction to revert}

\item{organization}{Name of your Briq organization}

\item{api_token}{Briq API token}
}
\description{
Revert a 'Briq' transaction
}
\examples{
\dontrun{
# Read organization and api_token from .Renviron file
bq_revert_transaction(transaction_id = "a1b-2c3")

# Manually enter in organization and api_token
bq_revert_transaction(
    transaction_id = "a1b-2c3", organization = "My Org", api_token = "xYz123"
)
}
}
