% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_groups.R
\name{bq_groups}
\alias{bq_groups}
\title{Briq user groups}
\usage{
bq_groups(organization = Sys.getenv("organization_name"),
  api_token = Sys.getenv("briq_api_token"))
}
\arguments{
\item{organization}{Name of your Briq organization}

\item{api_token}{Briq API token}
}
\value{
Returns a tibble of Briq user groups of your organization
}
\description{
List all the 'Briq' user groups of your organization
}
\examples{
\dontrun{
# Read organization and api_token from .Renviron file
bq_groups()

# Manually enter in organization and api_token
bq_groups(organization = "My Org", api_token = "xYz123")
}
}
