% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_users.R
\name{bq_users}
\alias{bq_users}
\title{Briq users}
\usage{
bq_users(organization = Sys.getenv("organization_name"),
  api_token = Sys.getenv("briq_api_token"))
}
\arguments{
\item{organization}{Name of your Briq organization}

\item{api_token}{Briq API token}
}
\value{
Returns a tibble of Briq users of your organization
}
\description{
List all the 'Briq' users of your organization
}
\examples{
\dontrun{
# Read organization and api_token from .Renviron file
bq_users()

# Manually enter in organization and api_token
bq_users(organization = "My Org", api_token = "xYz123")
}
}
